requirements.txt
numpy~=1.21.2
matplotlib~=3.4.3
scipy
setuptools~=58.2.0
pandas~=1.3.4
pyyaml
types-PyYAML